pyproject.toml
[tool.poetry]
name = "libraryapp"
version = "0.1.0"
description = "LibraryApp"
authors = ["Lukasz Ciesluk"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13"
flask = ">=3.1.1,<4.0.0"
flask-sqlalchemy = ">=3.1.1,<4.0.0"
pandas = ">=2.3.0,<3.0.0"
flasgger = ">=0.9.7.1,<0.10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
mypy = "^1.16.1"
pandas-stubs = "^2.2.3.250527"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
